fix: token cache max age expects ms but got seconds (#15)
there is a small hint at the end of `axios-token-interceptor`'s README.md
saying that the `getMaxAge` method of the tokenCache expects milliseconds but
authorization server return `expires_in` in seconds.
Therefore it is necessary to convert it, otherwise the cache will expire too fast.

The existing test worked, cause the execution was too fast. Adding a small delay
and executing the test again, shows that the token is not cached correctly.